Changing baseProducts to Product objects rather than strings

+++ app/src/main/java/com/elderj/qmarket/Product.java
@@ -3,10 +3,10 @@ package com.elderj.qmarket;
 public class Product {
 
     public final String name;
-    public final long buyPrice;
-    public final long sellPrice;
+    public final double buyPrice;
+    public final double sellPrice;
 
-    public Product(String name, long buyPrice, long sellPrice) {
+    public Product(String name, double buyPrice, double sellPrice) {
         this.name = name;
         this.buyPrice = buyPrice;
         this.sellPrice = sellPrice;
